refactor(infra-monitoring): keep the container helpers with the container entity
The status and readiness colours, labels and count builders are read by the
Containers table alone, so they sit beside it rather than in the module every
k8s entity imports.
